Add support for git https credentials
This implements #21. It is a simple way to provide git credentials not only for
ssh via keys, but also for https. In corporate settings with internet access
through proxies, ssh is often not allowed and one is forced to use https.

There are two ways to use this, either you provide a credential store and mount
it when running `vault`.

```
docker run -p 172.17.0.1:14242:3000 -v $PWD/store:/vault/store vault
```

Alternatively you can also start the vault with an empty store and interactively
add / remove credentials.

```
docker run -p 172.17.0.1:14242:3000 --name vault vault

# the following prompts you for a username and password which are stored in the
# running container
docker exec -it vault credentials set github.com
```

The usage inside a `Dockerfile` is the same as for the ssh key. Simply prepend
your `git clone` or other commands with `ONVAULT`.

The changes are really fairly simple:

 - One route is added in `index.js` to download the file with the credentials.
 - ONVAULT is extended to download the additional file and configure git, as
   well as revert the changes after executing the passed command.
 - The `credential` helper script is added to the container. This is actually
   not essential, but more of a convenience such that you don't need to create
   the credential files manually.

I did not yet update the README, but I can do that before merge if you are
willing to include this addition.

closes #21